Fix: Update PokemonSpeciesReadDto to include additional fields and change multiplier type in TypeEffectivenessReadDto

+++ src/main/java/com/poketeambuilder/dtos/front/type/effectiveness/TypeEffectivenessReadDto.java
@@ -1,7 +1,9 @@
 package com.poketeambuilder.dtos.front.type.effectiveness;
 
+import java.math.BigDecimal;
+
 import com.poketeambuilder.dtos.front.type.typing.TypeReadDto;
 
-public record TypeEffectivenessReadDto(TypeReadDto attackingType, TypeReadDto defendingType, double multiplier) {
+public record TypeEffectivenessReadDto(TypeReadDto attackingType, TypeReadDto defendingType, BigDecimal multiplier) {
     
 }
